WIP: AVX2 working, some tests fail

+++ libs/vectorise/benchmarks/parallel_dispatcher/kernel_bench.cpp
@@ -47,8 +47,8 @@ protected:
   {}
 
   ndarray_type a_, b_;
-};
-BENCHMARK_TEMPLATE_F(ParallelDispatcherKernelBench, kernel_implementation, float)
+}; 
+BENCHMARK_TEMPLATE_F(ParallelDispatcherKernelBench, kernel_implementation, double)
 (benchmark::State &st)
 {
   // Standard implementation
